+++ docs/conf.py
@@ -1,19 +1,22 @@
-"""
-Configuration file for the Sphinx documentation builder.
-"""
+# Configuration file for the Sphinx documentation builder.
+#
+# This file does only contain a selection of the most common options. For a
+# full list see the documentation:
+# http://www.sphinx-doc.org/en/master/config
 
 import os
-from datetime import datetime
+import datetime
 from pathlib import Path
+
+from packaging.version import Version
 
 import pyvista
 from pyvista.plotting.utilities.sphinx_gallery import DynamicScraper
 from sunpy_sphinx_theme import PNG_ICON
 
-from sunkit_pyvista import __version__
-
 # -- Read the Docs Specific Configuration --------------------------------------
-# This needs to be done before sunpy is imported
+
+# This needs to be done before sunkit-pyvista is imported
 on_rtd = os.environ.get("READTHEDOCS", None) == "True"
 if on_rtd:
     os.environ["SUNPY_CONFIGDIR"] = "/home/docs/"
@@ -24,17 +27,34 @@ if on_rtd:
     os.environ["PYDEVD_DISABLE_FILE_VALIDATION"] = "1"
 
 # -- Project information -----------------------------------------------------
+
+# The full version, including alpha/beta/rc tags
+from sunkit_pyvista import __version__
+
+_version = Version(__version__)
+version = release = str(_version)
+# Avoid "post" appearing in version string in rendered docs
+if _version.is_postrelease:
+    version = release = _version.base_version
+# Avoid long githashes in rendered Sphinx docs
+elif _version.is_devrelease:
+    version = release = f"{_version.base_version}.dev{_version.dev}"
+is_development = _version.is_devrelease
+is_release = not (_version.is_prerelease or _version.is_devrelease)
+
 project = "sunkit-pyvista"
-author = "SunPy Community"
-copyright = f"{datetime.now().year}, {author}"  # NOQA: A001, DTZ005
-release = __version__
+author = "The SunPy Community"
+copyright = f"{datetime.datetime.now().year}, {author}"  # noqa: A001
 
 # -- General configuration ---------------------------------------------------
+
+# Wrap large function/method signatures
+maximum_signature_line_length = 80
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named "sphinx.ext.*") or your custom
+# ones.
 extensions = [
-    "sphinx_automodapi.automodapi",
-    "sphinx_automodapi.smart_resolver",
-    "sphinx_changelog",
-    "sphinx_gallery.gen_gallery",
     "pyvista.ext.viewer_directive",
     "sphinx.ext.autodoc",
     "sphinx.ext.coverage",
@@ -45,34 +65,32 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.todo",
     "sphinx.ext.viewcode",
+    "sphinx_automodapi.automodapi",
+    "sphinx_automodapi.smart_resolver",
+    "sphinx_changelog",
     "sphinx_design",
+    "sphinx_gallery.gen_gallery",
 ]
 
-html_theme = "sunpy"
+# Add any paths that contain templates here, relative to this directory.
+# templates_path = ["_templates"]
 
-# For the linkcheck
-linkcheck_ignore = [
-    r"https://doi.org/\d+",
-    r"https://element.io/\d+",
-    r"https://github.com/\d+",
-    r"https://docs.sunpy.org/\d+",
-]
-linkcheck_anchors = False
-
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
-source_suffix = ".rst"
+
+# The suffix(es) of source filenames.
+source_suffix = {".rst": "restructuredtext"}
+
+# The master toctree document.
 master_doc = "index"
-nitpicky = True
-nitpick_ignore = []
-with Path("nitpick-exceptions.txt").open() as f:
-    for line in f:
-        if line.strip() == "" or line.startswith("#"):
-            continue
-        dtype, target = line.split(None, 1)
-        target = target.strip()
-        nitpick_ignore.append((dtype, target))
+
+# Treat everything in single ` as a Python reference.
+default_role = "py:obj"
 
 # -- Options for intersphinx extension ---------------------------------------
+
 intersphinx_mapping = {
     "python": (
         "https://docs.python.org/3/",
@@ -88,11 +106,65 @@ intersphinx_mapping = {
     ),
     "astropy": ("https://docs.astropy.org/en/stable/", None),
     "sunpy": ("https://docs.sunpy.org/en/stable", None),
-    "pyvista": ("https://docs.pyvista.org/", None),
+    "pyvista": ("https://docs.pyvista.org/version/stable", None),
     "sunkit_magex": ("https://docs.sunpy.org/projects/sunkit-magex/en/stable/", None),
 }
 
+# -- Options for HTML output -------------------------------------------------
+
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+html_theme = "sunpy"
+
+# Render inheritance diagrams in SVG
+graphviz_output_format = "svg"
+
+graphviz_dot_args = [
+    "-Nfontsize=10",
+    "-Nfontname=Helvetica Neue, Helvetica, Arial, sans-serif",
+    "-Efontsize=10",
+    "-Efontname=Helvetica Neue, Helvetica, Arial, sans-serif",
+    "-Gfontsize=10",
+    "-Gfontname=Helvetica Neue, Helvetica, Arial, sans-serif",
+]
+
+# Add any paths that contain custom static files (such as style sheets) here,
+# relative to this directory. They are copied after the builtin static files,
+# so a file named "default.css" will overwrite the builtin "default.css".
+# html_static_path = ["_static"]
+
+# By default, when rendering docstrings for classes, sphinx.ext.autodoc will
+# make docs with the class-level docstring and the class-method docstrings,
+# but not the __init__ docstring, which often contains the parameters to
+# class constructors across the scientific Python ecosystem. The option below
+# will append the __init__ docstring to the class-level docstring when rendering
+# the docs. For more options, see:
+# https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#confval-autoclass_content
+autoclass_content = "both"
+
+# -- Other options ----------------------------------------------------------
+
+# for the lint check
+lintcheck_ignore = [
+    r"https://doi.org/\d+",
+    r"https://element.io/\d+",
+    r"https://github.com/\d+",
+    r"https://docs.sunpy.org/\d+",
+]
+linkcheck_anchors = False
+
+nitpicky = True
+nitpick_ignore = []
+with Path("nitpick-exceptions.txt").open() as f:
+    for line in f:
+        if line.strip() == "" or line.startswith("#"):
+            continue
+        dtype, target = line.split(None, 1)
+        target = target.strip()
+        nitpick_ignore.append((dtype, target))
+
 # -- pyvista configuration ---------------------------------------------------
+
 os.environ["PYVISTA_BUILDING_GALLERY"] = "True"
 pyvista.BUILDING_GALLERY = True
 pyvista.global_theme.font.label_size = 18
@@ -107,6 +179,7 @@ if on_rtd or os.environ.get("CI"):
     pyvista.start_xvfb()
 
 # -- Sphinx Gallery ------------------------------------------------------------
+
 sphinx_gallery_conf = {
     "backreferences_dir": Path("generated") / "modules",
     "filename_pattern": "^((?!skip_).)*$",