+++ examples/floating_sphere.py
@@ -7,10 +7,12 @@ This example demonstrates how to take a collection of points which you want to
 visualize as either a cloud of points or a surface.
 """
 
-import astropy.units as u
 import numpy as np
-import sunpy.map
+
+import astropy.units as u
 from astropy.coordinates import SkyCoord, SphericalRepresentation
+
+import sunpy.map
 from sunpy.sun.constants import radius as Rsun
 
 from sunkit_pyvista import SunpyPlotter
@@ -65,7 +67,7 @@ plotter.plot_map(aia)
 # sphinx_gallery_defer_figures
 
 cloud = plotter.coordinates_to_polydata(point_cloud)
-_ = plotter.plotter.add_points(cloud, point_size=0.7, color="cyan", style="points_gaussian")
+_ = plotter.add_points(cloud, point_size=0.7, color="cyan", style="points_gaussian")
 
 ################################################################################
 # Next we want to build a surface from these points.
@@ -75,7 +77,7 @@ _ = plotter.plotter.add_points(cloud, point_size=0.7, color="cyan", style="point
 # sphinx_gallery_defer_figures
 
 surf = cloud.delaunay_3d()
-_ = plotter.plotter.add_mesh(surf)
+_ = plotter.add_mesh(surf)
 
 ################################################################################
 # Finally set up the camera position and focus.