+++ sunkit_pyvista/plotter.py
@@ -1,16 +1,19 @@
 import contextlib
 from pathlib import Path
 
-import astropy.units as u
 import numpy as np
 import pyvista as pv
+from matplotlib import colors
+
+import astropy.units as u
 from astropy.constants import R_sun
 from astropy.coordinates import Longitude, SkyCoord
 from astropy.visualization import AsymmetricPercentileInterval
 from astropy.visualization.wcsaxes import Quadrangle
-from matplotlib import colors
+
 from sunkit_magex.pfss.coords import strum2cart
-from sunpy.coordinates import HeliocentricInertial, Helioprojective
+from sunpy.coordinates import HeliocentricInertial
+from sunpy.coordinates.screens import SphericalScreen
 from sunpy.coordinates.utils import get_rectangle_coordinates
 from sunpy.map.maputils import all_corner_coords_from_map
 
@@ -19,24 +22,24 @@ from sunkit_pyvista.utils import get_limb_coordinates
 __all__ = ["SunpyPlotter"]
 
 
-class SunpyPlotter:
+class SunpyPlotter(pv.Plotter):
     """
     A plotter for 3D data.
 
-    This class wraps `pyvsita.Plotter` to provide coordinate-aware plotting.
+    This class inherits `pyvista.Plotter` so we can provide coordinate-aware plotting.
     For now, all coordinates are converted to
     a specific frame (`~sunpy.coordinates.HeliocentricInertial` by default),
     and distance units are such that :math:`R_{sun} = 1`.
 
     Parameters
     ----------
-    coordinate_frame : `astropy.coordinates.BaseFrame`
+    coordinate_frame : `astropy.coordinates.BaseCoordinateFrame`
         Coordinate frame of the plot. The x, y, z axes of the pyvista plotter
         will be the x, y, z axes in this coordinate system.
     obstime : `astropy.time.Time`
         The obstime to use for the default coordinate frame if
-        `coordinate_frame=` is not specified.  Must not be specified if
-        `coordinate_frame` is given.
+        ``coordinate_frame=`` is not specified.  Must not be specified if
+        ``coordinate_frame`` is given.
     kwargs : dict
         All other keyword arguments are passed through to `pyvista.Plotter`.
 
@@ -46,15 +49,14 @@ class SunpyPlotter:
         Stores a reference to all the plotted meshes in a dictionary.
     """
 
-    def __init__(self, *, coordinate_frame=None, obstime=None, **kwargs):
+    def __init__(self, *args, coordinate_frame=None, obstime=None, **kwargs):
+        super().__init__(*args, **kwargs)
         if coordinate_frame is not None and obstime is not None:
             msg = "Only coordinate_frame or obstime can be specified, not both."
             raise ValueError(msg)
         if coordinate_frame is None:
             coordinate_frame = HeliocentricInertial(obstime=obstime)
         self._coordinate_frame = coordinate_frame
-        self._plotter = pv.Plotter(**kwargs)
-        self.camera = self._plotter.camera
         self.all_meshes = {}
 
     @property
@@ -63,21 +65,6 @@ class SunpyPlotter:
         Coordinate frame of the plot.
         """
         return self._coordinate_frame
-
-    @property
-    def plotter(self):
-        """
-        `pyvista.Plotter`.
-        """
-        return self._plotter
-
-    def show(self, *args, **kwargs):
-        """
-        Show the plot.
-
-        See `pyvista.Plotter.show` for accepted arguments.
-        """
-        self.plotter.show(*args, **kwargs)
 
     def _extract_color(self, mesh_kwargs):
         """
@@ -142,8 +129,8 @@ class SunpyPlotter:
 
     def coordinates_to_polydata(self, coords):
         """
-        Convert a set of coordinates in a `.SkyCoord` to a `pyvista.PolyData`
-        mesh.
+        Convert a set of coordinates in a `~astropy.coordinates.SkyCoord` to a
+        `pyvista.PolyData` mesh.
 
         Parameters
         ----------
@@ -170,7 +157,7 @@ class SunpyPlotter:
         """
         camera_position = self._coords_to_xyz(coord)
         pos = tuple(camera_position)
-        self.plotter.camera.position = pos
+        self.camera.position = pos
 
     def set_camera_focus(self, coord):
         """
@@ -183,7 +170,7 @@ class SunpyPlotter:
         """
         camera_position = self._coords_to_xyz(coord)
         pos = tuple(camera_position)
-        self.plotter.set_focus(pos)
+        self.set_focus(pos)
 
     @u.quantity_input
     def set_view_angle(self, angle: u.deg):
@@ -197,10 +184,10 @@ class SunpyPlotter:
         """
         view_angle = angle.to_value(u.deg)
         if not (view_angle > 0 and view_angle <= 180):
-            msg = "specified view angle must be " "0 deg < angle <= 180 deg"
+            msg = "specified view angle must be 0 deg < angle <= 180 deg"
             raise ValueError(msg)
         zoom_value = self.camera.view_angle / view_angle
-        self.plotter.camera.zoom(zoom_value)
+        self.camera.zoom(zoom_value)
 
     def _map_to_mesh(self, m, *, assume_spherical=True):
         """
@@ -220,7 +207,7 @@ class SunpyPlotter:
         corner_coords = all_corner_coords_from_map(m)
 
         if assume_spherical:
-            context = Helioprojective.assume_spherical_screen(
+            context = SphericalScreen(
                 m.observer_coordinate,
                 only_off_disk=True,
             )
@@ -292,7 +279,7 @@ class SunpyPlotter:
             interval bounded by the two numbers.
         assume_spherical_screen : bool, optional
             If `True` (default) then off-limb pixels are plotted using
-            :meth:`sunpy.coordinates.Helioprojective.assume_spherical_screen`.
+            `sunpy.coordinates.screens.SphericalScreen`.
             If `False`, off-limb pixels are not plotted.
         **kwargs :
             Keyword arguments are handed to `pyvista.Plotter.add_mesh`.
@@ -305,7 +292,7 @@ class SunpyPlotter:
                     clip_interval=clip_interval,
                 )
             else:
-                msg = "Clip percentile interval must be " "specified as two numbers."
+                msg = "Clip percentile interval must be specified as two numbers."
                 raise ValueError(
                     msg,
                 )
@@ -313,7 +300,7 @@ class SunpyPlotter:
             clim = [0, 1]
         cmap = self._get_cmap(kwargs, m)
         kwargs.setdefault("show_scalar_bar", False)
-        self.plotter.add_mesh(map_mesh, cmap=cmap, clim=clim, **kwargs)
+        self.add_mesh(map_mesh, cmap=cmap, clim=clim, **kwargs)
         map_mesh.add_field_data([cmap], "cmap")
         self._add_mesh_to_dict(block_name="maps", mesh=map_mesh)
 
@@ -361,7 +348,7 @@ class SunpyPlotter:
         point_mesh.add_field_data(color, "color")
 
         kwargs["render_lines_as_tubes"] = kwargs.pop("render_lines_as_tubes", True)
-        self.plotter.add_mesh(point_mesh, color=color, smooth_shading=True, **kwargs)
+        self.add_mesh(point_mesh, color=color, smooth_shading=True, **kwargs)
         self._add_mesh_to_dict(block_name="coordinates", mesh=point_mesh)
 
     def plot_solar_axis(self, *, length=2.5, arrow_kwargs=None, **kwargs):
@@ -391,7 +378,7 @@ class SunpyPlotter:
         )
         color = self._extract_color(kwargs)
         arrow_mesh.add_field_data(color, "color")
-        self.plotter.add_mesh(arrow_mesh, color=color, **kwargs)
+        self.add_mesh(arrow_mesh, color=color, **kwargs)
         self._add_mesh_to_dict(block_name="solar_axis", mesh=arrow_mesh)
 
     def plot_quadrangle(
@@ -456,7 +443,7 @@ class SunpyPlotter:
         quad_block = quad_block.tube(radius=radius)
         color = self._extract_color(kwargs)
         quad_block.add_field_data(color, "color")
-        self.plotter.add_mesh(quad_block, color=color, **kwargs)
+        self.add_mesh(quad_block, color=color, **kwargs)
         self._add_mesh_to_dict(block_name="quadrangles", mesh=quad_block)
 
     def plot_field_lines(self, field_lines, *, color_func=None, **kwargs):
@@ -505,7 +492,7 @@ class SunpyPlotter:
 
             kwargs["render_lines_as_tubes"] = kwargs.pop("render_lines_as_tubes", True)
             kwargs["line_width"] = kwargs.pop("line_width", 2)
-            self.plotter.add_mesh(spline, color=color, **kwargs)
+            self.add_mesh(spline, color=color, **kwargs)
             field_line_meshes.append(spline)
 
         self._add_mesh_to_dict(block_name="field_lines", mesh=spline)
@@ -532,7 +519,7 @@ class SunpyPlotter:
         pfss_out_pv = pv.StructuredGrid(x_arr, y_arr, z_arr)
         pfss_out_pv["Br"] = bc_r.ravel("F")
         isos_br = pfss_out_pv.contour(isosurfaces=1, rng=[0, 0])
-        self.plotter.add_mesh(isos_br, **kwargs)
+        self.add_mesh(isos_br, **kwargs)
         self._add_mesh_to_dict(block_name="current_sheet", mesh=isos_br)
 
     def save(self, filepath, *, overwrite=False):
@@ -583,7 +570,7 @@ class SunpyPlotter:
             else:
                 color = dict(block.field_data).get("color", None)
                 cmap = dict(block.field_data).get("cmap", [None])[0]
-                self.plotter.add_mesh(block, color=color, cmap=cmap)
+                self.add_mesh(block, color=color, cmap=cmap)
 
     def load(self, filepath):
         """
@@ -622,5 +609,5 @@ class SunpyPlotter:
         color = self._extract_color(mesh_kwargs=kwargs)
         limb_block = limb_block.tube(radius=radius)
         limb_block.add_field_data(color, "color")
-        self.plotter.add_mesh(limb_block, color=color, **kwargs)
+        self.add_mesh(limb_block, color=color, **kwargs)
         self._add_mesh_to_dict(block_name="limbs", mesh=limb_block)