+++ sunkit_pyvista/tests/test_magex.py
@@ -2,15 +2,20 @@
 This file contains tests for any methods that use sunkit-magex.
 """
 
-import astropy.constants as const
-import astropy.units as u
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 import pyvista as pv
-import sunpy.map as smap
-from astropy.coordinates import SkyCoord
 from matplotlib import colors
+
+import astropy.constants as const
+import astropy.units as u
+from astropy.coordinates import SkyCoord
+
+import sunpy.map as smap
+from sunkit_magex import pfss
+from sunkit_magex.pfss import tracing
+from sunkit_magex.pfss.sample_data import get_gong_map
 
 from sunkit_pyvista import SunpyPlotter
 
@@ -18,11 +23,6 @@ pv.OFF_SCREEN = True
 
 
 def test_field_lines_figure(aia171_test_map, plotter, verify_cache_image):
-    pfss = pytest.importorskip("sunkit_magex.pfss")
-
-    from sunkit_magex.pfss import tracing
-    from sunkit_magex.pfss.sample_data import get_gong_map
-
     gong_fname = get_gong_map()
     gong_map = smap.Map(gong_fname)
     nrho = 35
@@ -32,7 +32,7 @@ def test_field_lines_figure(aia171_test_map, plotter, verify_cache_image):
     lat, lon = np.meshgrid(lat, lon, indexing="ij")
     lat, lon = lat.ravel() * u.rad, lon.ravel() * u.rad
     radius = 1.2
-    tracer = tracing.FortranTracer()
+    tracer = tracing.PythonTracer()
     input_ = pfss.Input(gong_map, nrho, rss)
     output_ = pfss.pfss(input_)
     seeds = SkyCoord(lon, lat, radius * const.R_sun, frame=gong_map.coordinate_frame)
@@ -63,7 +63,7 @@ def test_field_lines_and_color_func(plotter, verify_cache_image):
     lat, lon = np.meshgrid(lat, lon, indexing="ij")
     lat, lon = lat.ravel() * u.rad, lon.ravel() * u.rad
     radius = 1.2
-    tracer = tracing.FortranTracer()
+    tracer = tracing.PythonTracer()
     input_ = pfss.Input(gong_map, nrho, rss)
     output_ = pfss.pfss(input_)
     seeds = SkyCoord(lon, lat, radius * const.R_sun, frame=gong_map.coordinate_frame)